requirements.txt
Django == 3.2
django-fernet-fields == 0.6